+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ### Features
 - Added Lotus Atoll to RFU Party Finder (Only appears to people on this version onwards)
+- Added Lotus Atoll SCs
 
 ### Changes
 - Improved Sea Creature configuration error handling to prevent malformed JSON entries from breaking the loading process.
+- Removed normal scs from bayou
+
+### Back-end
+- Added an option to disable automatic Sea Creature configuration synchronization in Backend Settings.
 
 ### Fixes
 - Fixed hotspots height calculation to correctly handle waterlogged blocks.